Slight code restructure and fix shared ARView
