+++ InvisibleMap2/ContentView.swift
@@ -2,7 +2,7 @@
 //  ContentView.swift
 //  InvisibleMap2
 //
-//  Created by tad on 9/15/21.
+//  Created by Allison Li and Ben Morris on 9/15/21.
 //  Copyright © 2021 Occam Lab. All rights reserved.
 //
