+++ InvisibleMap2/State Machine/Controllers/MapNavigator.swift
@@ -1,0 +1,9 @@
+//
+//  MapNavigator.swift
+//  InvisibleMap2
+//
+//  Created by Allison Li and Ben Morris on 9/24/21.
+//  Copyright © 2021 Occam Lab. All rights reserved.
+//
+
+import Foundation
